requirements.txt
pandas>=2.0
numpy>=1.24
scipy>=1.10
pyarrow>=12.0
requests>=2.28
python-dotenv>=1.0
pyyaml>=6.0
pytz>=2023.3
python-telegram-bot>=20.0
